+++ src/main/java/org/folio/orders/lines/update/instance/WithoutHoldingOrderLineUpdateInstanceStrategy.java
@@ -34,7 +34,7 @@ public class WithoutHoldingOrderLineUpdateInstanceStrategy implements OrderLineU
 
     return new DBClient(rqContext.getContext(), rqContext.getHeaders()).getPgClient()
       .withTrans(conn -> titleService.updateTitle(storagePol, instanceId, conn)
-        .compose(poLine -> poLinesService.updateInstanceIdForPoLine(poLine, holder.instance(), conn)))
+        .compose(poLine -> poLinesService.updateInstanceIdForPoLine(poLine, holder.instance(), conn, rqContext.getHeaders())))
       .onSuccess(v -> log.info("updateInstance:: Instance was updated successfully, poLine id={}", storagePol.getId()))
       .onFailure(err -> log.warn("updateInstance:: Instance failed to update, poLine id={}", storagePol.getId(), err))
       .mapEmpty();