+++ src/test/java/org/folio/orders/lines/update/OrderLineUpdateInstanceHandlerTest.java
@@ -34,6 +34,7 @@ import org.apache.logging.log4j.Logger;
 import org.folio.dao.PostgresClientFactory;
 import org.folio.dao.lines.PoLinesDAO;
 import org.folio.dao.lines.PoLinesPostgresDAO;
+import org.folio.repository.CustomFieldsRepository;
 import org.folio.event.service.AuditOutboxService;
 import org.folio.orders.lines.update.instance.WithHoldingOrderLineUpdateInstanceStrategy;
 import org.folio.orders.lines.update.instance.WithoutHoldingOrderLineUpdateInstanceStrategy;
@@ -588,6 +589,11 @@ public class OrderLineUpdateInstanceHandlerTest extends TestBase {
   }
 
    static class ContextConfiguration {
+     @Bean
+     CustomFieldsRepository customFieldsRepository() {
+       return mock(CustomFieldsRepository.class);
+     }
+
      @Bean
      PoLinesDAO poLinesDAO() {
        return new PoLinesPostgresDAO();